refactor(procedures): Rename FormFillingStatus
