+++ packages/cozy-procedures/src/components/CompletedFromMyselfStatus.spec.jsx
@@ -1,22 +1,22 @@
 import React from 'react'
 
-import { FormFillingStatus } from './FormFillingStatus'
+import { CompletedFromMyselfStatus } from './CompletedFromMyselfStatus'
 
 const fakeT = (msgId, args) => {
   return `${msgId} ${JSON.stringify(args)}`
 }
 
-describe('FormFillingStatus component', () => {
+describe('CompletedFromMyselfStatus component', () => {
   it('should display nothing if no data is filled', () => {
     const component = shallow(
-      <FormFillingStatus completed={0} total={10} t={fakeT} />
+      <CompletedFromMyselfStatus completed={0} total={10} t={fakeT} />
     ).getElement()
     expect(component).toBeNull()
   })
 
   it('should display a status if there are some completed data', () => {
     const component = mount(
-      <FormFillingStatus completed={5} total={10} t={fakeT} />
+      <CompletedFromMyselfStatus completed={5} total={10} t={fakeT} />
     )
     expect(component).toMatchSnapshot()
   })
